requirements.txt
PyYAML
numpy
torch
pytest
psutil
onnxruntime
matplotlib
seaborn
gym
faiss-cpu
scipy
# faiss-gpu